move util.ts into rest folder
